split out config environment file
